EA: Fix compilation with gcc-4.5.2

+++ src/SMESH/SMESH_MeshEditor.cxx
@@ -11157,7 +11157,7 @@ bool SMESH_MeshEditor::DoubleNodesOnGroupBoundaries( const std::vector<TIDSorted
         {
           vector<int> nodes = ite->first;
           vector<int> orderDom = ite->second;
-          vector<int> orderedNodes;
+          vector<long long int> orderedNodes;
           if (nodes.size() == 2)
             {
               //MESSAGE(" use edgesMultiDomains " << nodes[0] << " " << nodes[1]);
